Add actuator handler app and docker image
- Add a new Docker image which contains a python 'actuator' application,
which can be used as an example file that implements some activateable
mechanism (possibly outside the camera idiom's bounds)
  For example, the actuator class could be extended to drive an electric
  motor to open a door. Or it could be extended to play some sound over
  a speaker or piezo siren.
- Start both docker images using a single docker-compose.yml file, and
have the guest processes talk to each other (configured by the yaml file).
- Move cloudtalker into its own src/ subdirectory, and create a new
subdir for the actuator object. Other subdirectories may be added in a
similar fashion as the only thing that's needed is an updated
docker-compose.yml file.